Begin work on auto-sync-enforcing, add git magic
Shallow clonging requires --no-single-branch to
show all remote branches with git branch -a .
This is required, REQUIRED for portage to be
able to enforce the values in the repos.conf
config files.

auto-sync-enforcing=yes is meant to be able to enforce
both sync-uri and sync-branch values to match the git
repository ones by mangling the git repository into shape.